% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_param.R
\name{beta_param}
\alias{beta_param}
\title{Estimate Beta Parameter}
\usage{
beta_param(mean, sd)
}
\arguments{
\item{mean}{Numeric. The mean.}

\item{sd}{Numeric. The standard deviation.}
}
\description{
Estimates beta parameter given mean (rates) and SD for the
the "untruncated" beta distribution.
}
\details{
Updated using analytical solution from Kyle Wilson.
}
\examples{
\dontrun{
library(JoeModelCE)
JoeModelCE::beta_param(0.8, 0.2)$beta
}

}
