% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_CumulativeEffect.R
\name{beta_stretch_val}
\alias{beta_stretch_val}
\title{beta_stretch_val}
\usage{
beta_stretch_val(mn, variance = NA, sigma = NA)
}
\description{
Calculate parameters for a stretched beta distribution.
}
\keyword{internal}
