% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_JoeModel.R
\name{ce.func}
\alias{ce.func}
\title{Apply CSC Across dataframe}
\usage{
ce.func(df)
}
\arguments{
\item{df}{A dataframe}
}
\description{
Apply CSC Across dataframe
}
\details{
ddply function to calculate cumulative system capacity, the df contains system capacity for each stressor for a given HUC and simulation
}
\keyword{internal}
