% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_CumulativeEffect.R
\name{d.vec.bh}
\alias{d.vec.bh}
\title{d.vec.bh}
\usage{
d.vec.bh(df, Ks, N, bh_dd_stages)
}
\description{
create vector of density dependence effects using compensation ratios
}
\keyword{internal}
