% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_model_bh_dd.R
\name{dd.N.bh}
\alias{dd.N.bh}
\title{dd.N.bh}
\usage{
dd.N.bh(dat = NA, t = NA, st = NA, N = NA, N_prev = NA, bh_dd_stages = NA)
}
\description{
Apply location and stage-specific Beverton-Holt density dependent constraints
}
\keyword{internal}
