% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationModel_Utility.R
\name{esub}
\alias{esub}
\title{Utility Function 1}
\usage{
esub(expr, sublist)
}
\arguments{
\item{expr}{A number}

\item{sublist}{A number}
}
\description{
Utility Function 1
}
\keyword{internal}
