% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_CumulativeEffect.R
\name{init_pop}
\alias{init_pop}
\title{init_pop}
\usage{
init_pop(mx, Nt, p.rep)
}
\arguments{
\item{mx}{Population Matrix}

\item{Nt}{Desire adult density}

\item{p.rep}{maturity vector - 1: mature; 0: Juvenle; length Tmax + 1}
}
\description{
init_pop
}
\keyword{internal}
