% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_matrix_load.R
\name{interaction_matrix_load}
\alias{interaction_matrix_load}
\title{Load Interaction Matrices}
\usage{
interaction_matrix_load(mint_names = NA, filename = NA)
}
\arguments{
\item{mint_names}{Character vector of interaction matrix names.}

\item{filename}{Character string of Excel file.}
}
\value{
a list object of interaction matrix objects.
}
\description{
Load interaction Excel Matrices
}
\details{
Load 2-factor interaction matrices from
Excel workbook (if they exist). This function is run from
within StressorResponseWorkbook().
}
