% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_matrix_sys_cap.R
\name{interaction_matrix_sys_cap}
\alias{interaction_matrix_sys_cap}
\title{Interaction Matrix System Capacity}
\usage{
interaction_matrix_sys_cap(MInt, sc.dose.df, adult_sys_cap)
}
\arguments{
\item{MInt}{One item from object returned from StressorResponseWorkbook()}

\item{sc.dose.df}{Data frame with columns HUC, Stressor, simulation, dose
and sys.cap.}

\item{adult_sys_cap}{Should the Joe Model be run only with variables
identified for adult system capacity.}
}
\description{
Use run a 2D linear interpolation with interaction matrix to
calculate system capacity.
}
\details{
Runs through an individual MInt_ worksheet object supplied from
sr_wb_dat$MInt and calculates system capacity. interaction_matrix_sys_cap()
should be run with apply on sr_wb_dat$MInt. Returns sc.dose.df with updated
system capacity and filtered for variables.
}
