% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_Response.R
\name{mean_Response}
\alias{mean_Response}
\title{Mean Response Function}
\usage{
mean_Response(n.stressors = NA, str.list = NA, main = NA)
}
\arguments{
\item{n.stressors}{Numeric. The number of stressors.}

\item{str.list}{List object. A list object of dataframes of stressor response relationships.}

\item{main}{Dataframe. A dataframe of the Main cover sheet for the stressor response relationships .}
}
\description{
Generate the mean response function for each stressor.
}
\details{
Creates functions to interpolate the mean system capacity, SD, lower limit and upper limit.
}
