% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_JoeModel.R
\name{mn.sd.huc}
\alias{mn.sd.huc}
\title{Cumulative System Capacity Calculation}
\usage{
mn.sd.huc(df)
}
\arguments{
\item{df}{A dataframe}
}
\description{
Cumulative System Capacity Calculation
}
\details{
Function to calculate cumulative system capacity mean and sd across simulations for each HUC from a series of Monte Carlo sims
}
\keyword{internal}
