% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoeModelCE_Utility.R
\name{pmx_eval}
\alias{pmx_eval}
\title{Utility Function 3}
\usage{
pmx_eval(mx, vars, byrow = TRUE)
}
\arguments{
\item{mx}{A argument}

\item{vars}{A argument}

\item{byrow}{A argument}
}
\description{
Utility Function 3
}
\keyword{internal}
