% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_model_K_adj.R
\name{pop_model_K_adj}
\alias{pop_model_K_adj}
\title{K_adj}
\usage{
pop_model_K_adj(replicates = 100, dat, mx, dx, Nyears = 250, H.mx = NULL)
}
\arguments{
\item{replicates}{Number of replicates}

\item{dat}{Life history data}

\item{mx}{A projection matrix expression}

\item{dx}{A matrix of density-dependence effect}

\item{Nyears}{Years to run simulation}

\item{H.mx}{A harm projection matrix}
}
\value{
A K_adj object to make adjustments
}
\description{
Define K_adj value.
}
\details{
Initialize populations and find carrying capacity for each life stage.
Carrying capacity is defined at the adult stage and estimated for all other
life-stages using the stable-stage distribution (i.e., if the carrying
capacity of the adult stage was set to 100 the carrying capacity
for a stage below (stage x) is the amount of stage x necessary to result
in 100 adults at equilibrium conditions).
}
