% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_model_multi_huc.R
\name{pop_model_multi_huc}
\alias{pop_model_multi_huc}
\title{Population model across multiple HUCs}
\usage{
pop_model_multi_huc(
  dir_in,
  dir_out,
  dose_file = NA,
  sr_wb_dat = NA,
  life_cycle_params = NA,
  HUC_ID = NA,
  n_years = 100,
  MC_sims = 10,
  stressors = NA,
  output_type = "full"
)
}
\arguments{
\item{dir_in}{Character path of the directory where the JOEMODEL input files live. Make sure that  the directory holds: 1) stressor response Excel file; 2) stressor magnitude Excel file; 3) life cycle parameters Excel file}

\item{dir_out}{character path to the directory where new files will be saved}

\item{dose_file}{A character that specified the name of the Excel filename (with extension .xlsx) of the stressor Response file (e.g."stressor_response_fixed_ARTR.xlsx")}

\item{sr_wb_dat}{A character that specified the name of the Excel filename (with extension .xlsx) of the stressor magnitude file (e.g."stressor_magnitude_unc_ARTR.xlsx")}

\item{life_cycle_params}{A character that specified the name of the Excel filename (with extension .xlsx) that folds the life cycle parameters (e.g."life_cycles.csv")}

\item{HUC_ID}{a numeric vector that holds the IDs of the HUCs to be assessed}

\item{n_years}{A numeric number that specifies hoe long should the JOEMODEL be run.If not specified, it will default to 300}

\item{MC_sims}{A number that specified how many Monte Carlo (MC) simulations will be run for each HUC. Each MC run will sample from the stressor's distribution and return the corresponding mean system capacity. If not specified; it will default to 5}

\item{stressors}{A character vector that holds the stressors that you want to include when running the sensitivity analysis. These stressors are common to ALL HUCs}

\item{output_type}{A character that should be set either to adult or to "full". The former will only return the adult population. The latter will return the population for all stages}
}
\value{
The function will return:
\itemize{
\item Excel files that hold the population numbers. A separate file is created for each HUC. Within each file, the populations of either the 4 stages (if you opt for output_type="full") or the population of the adult phase (if you opt for output_type="adult) is saved over time (for both the baseline and the ce scenarios) for all the Monte Carlo simulations. The columns are named to indicate the population stage, the scenario (baseline or ce), and the monte carlo simulation number associated with the output (e.g. column Stage_1_baseline_sim_1 = Population of Stage 1 for the basline scenario under simulation 1).
\item The function also returns a list of all the lambdas calculated over time for the baseline and ce scenarios across the number of Monte Carlo simulations specified. Each list element is named by its corresponding HUC and holds the result for that particular HUC
}
\description{
Runs the PopulationModel_Run in batch mode across different HUCS and saved the population outputs in excel files and saves the lambdas in a list for both the baseline and the ce scenarios.
}
\examples{
\dontrun{
library(JoeModelCE)
POP <- PopulationModel_Batch_Run(dir_in="C:/JOEMODEL", dir_out="C:/JOEMODEL/Results", dose_file = "stressor_response_fixed_ARTR.xlsx",
                               sr_wb_dat = "stressor_magnitude_unc_ARTR.xlsx", life_cycle_params = "life_cycles.csv",
         HUC_ID = c(1701010201, 1701010202), n_years = 300, MC_sims = 5, stressors = NA, output_type="full")
        }

}
