% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoeModelCE_Utility.R
\name{proc}
\alias{proc}
\title{Utility Function 2}
\usage{
proc(e, env = parent.frame())
}
\arguments{
\item{e}{An expression}

\item{env}{parent frame}
}
\description{
Utility Function 2
}
\keyword{internal}
