% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projection_DD.R
\name{Projection_DD}
\alias{Projection_DD}
\title{Project matrix with Density Dependence}
\usage{
Projection_DD(
  M.mx = NA,
  D.mx = NULL,
  H.mx = NULL,
  dat = NA,
  Nyears = 100,
  K = NA,
  p.cat = NA,
  CE_df = NULL
)
}
\arguments{
\item{M.mx}{A projection matrix expression}

\item{D.mx}{A matrix of density-dependence effect}

\item{H.mx}{A harm projection matrix}

\item{dat}{Life history data}

\item{Nyears}{Years to run simulation}

\item{K}{The population carrying Capacity}

\item{p.cat}{Probability of catastrophic event.}

\item{CE_df}{TODO add description}
}
\description{
Project the matrix model forward in time with density dependence
}
\details{
TODO add details.
}
