% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_PopulationModel.R
\name{rbeta2}
\alias{rbeta2}
\title{rbeta equivelent from TruncatedDistributions package}
\usage{
rbeta2(n, shape1, shape2, ncp = 0)
}
\description{
rbeta equivelent from TruncatedDistributions package
}
\keyword{internal}
