% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_JoeModel.R
\name{rtnorm_TruncatedDistributions}
\alias{rtnorm_TruncatedDistributions}
\title{rtnorm from Truncated Distributions}
\usage{
rtnorm_TruncatedDistributions(n, mean = 0, sd = 1, a = -Inf, b = Inf)
}
\description{
rtnorm from Truncated Distributions
}
\details{
rtnorm from Truncated Distributions
}
\keyword{internal}
