% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StressorMagnitudeWorkbook.R
\name{StressorMagnitudeWorkbook}
\alias{StressorMagnitudeWorkbook}
\title{Get Stressor Magnitude Data from Excel Workbook}
\usage{
StressorMagnitudeWorkbook(filename = NA, scenario_worksheet = NA)
}
\arguments{
\item{filename}{A string. Relative file name to the stressor magnitude Excel workbook.}

\item{scenario_worksheet}{A string. Worksheet name for the scenario.}
}
\value{
A stressor magnitude list object for the JoeModelCE
}
\description{
Extract the stressor magnitude data from each sheet in the Excel workbook.
}
\details{
For more information about data format requirements, see the Data Upload section of the vignette.
}
