% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StressorResponseWorkbook.R
\name{StressorResponseWorkbook}
\alias{StressorResponseWorkbook}
\title{Get Stressor Response Data from Excel Workbook}
\usage{
StressorResponseWorkbook(filename = NA)
}
\arguments{
\item{filename}{A string. Relative file name to the stressor response Excel workbook.}
}
\value{
A list object of the main_sheet, stressor_names and stressor response dataframes (sr_dat)
}
\description{
Extract the stressor response data from each sheet in the Excel workbook.
}
\details{
TODO add details.
}
