% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoeModelCE_Utility.R
\name{sys.cap.func}
\alias{sys.cap.func}
\title{Apply CSC Across stressors}
\usage{
sys.cap.func(
  f.dose.df,
  f.main.df,
  f.stressor.df,
  f.mean.resp.list,
  n.sims = MC.sims
)
}
\arguments{
\item{df}{A dataframe}
}
\description{
Apply CSC Across stressors
}
\details{
function to calculate system capacity for each stressor.  Takes dataframes for doses and stressor.list plus a list for the approx functions (f. for local list or dataframe) note some stressors have multiple doses (Additive interaction)
}
\keyword{internal}
