% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemCapacity.R
\name{SystemCapacity}
\alias{SystemCapacity}
\title{System Capacity}
\usage{
SystemCapacity(
  f.dose.df,
  f.main.df,
  f.stressor.df,
  f.mean.resp.list,
  n.sims = 100
)
}
\arguments{
\item{df}{A dataframe}
}
\description{
Calculates the system capacity for each stressor.
}
\details{
function to calculate system capacity for each stressor.  Takes dataframes for doses and stressor.list plus a list for the approx functions (f. for local list or dataframe) note some stressors have multiple doses (Additive interaction)
}
