% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_JoeModel.R
\name{tbeta_rnd}
\alias{tbeta_rnd}
\title{tbeta Random}
\usage{
tbeta_rnd(mn_est, sd_est, low.limit = 0, up.limit = 1)
}
\arguments{
\item{mn_est}{Vector of mean values.}

\item{sd_est}{Vector of standard deviations.}

\item{low.limit}{Numeric. Defaults to 0.}

\item{up.limit}{Numeric. Defaults to 1 .}
}
\description{
Returns N random estimates for truncated beta distribution given vectors for mean and SD for the untruncated distribution of system capacity.
}
\details{
Function returns returns one random value for each mean and sd value. Will not run with 0 or 1 values.
}
\keyword{internal}
