% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_CumulativeEffect.R
\name{bh_stage}
\alias{bh_stage}
\title{bh_stage}
\usage{
bh_stage(cr = 2, k = 100, x = 1)
}
\description{
bh_stage
}
\details{
functions used to calculate density dependence and stage-specific
survival/transitions
}
\keyword{internal}
