% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_CumulativeEffect.R,
%   R/Utility_PopulationModel.R
\name{pmx_eval}
\alias{pmx_eval}
\title{pmx_eval}
\usage{
pmx_eval(mx, vars, byrow = TRUE)

pmx_eval(mx, vars, byrow = TRUE)
}
\description{
pmx_eval

Project Matrix Forward in Time
}
\keyword{internal}
