% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationModel_Utility.R
\name{pmx_eval}
\alias{pmx_eval}
\title{Project Matrix Forward in Time}
\usage{
pmx_eval(mx, vars, byrow = TRUE)
}
\arguments{
\item{mx}{A argument}

\item{vars}{A argument}

\item{byrow}{A argument}
}
\description{
Project Matrix Forward in Time
}
\keyword{internal}
