% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_model_matrix_elements.R
\name{pop_model_matrix_elements}
\alias{pop_model_matrix_elements}
\title{Population Model Matrix Elements}
\usage{
pop_model_matrix_elements(pop_mod_setup = NA)
}
\arguments{
\item{pop_mod_setup}{List. Object returned from population_model_setup().}
}
\value{
a list object of symbolic objects.
}
\description{
Build population model matrix elements.
}
\details{
TODO add details.
}
