% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_model_setup.R
\name{pop_model_setup}
\alias{pop_model_setup}
\title{Population Model Setup}
\usage{
pop_model_setup(life_cycles = NA)
}
\arguments{
\item{life_cycles}{Data frame. Raw life cycles.csv data frame.}
}
\value{
a list object of symbolic objects.
}
\description{
Generate symbolic objects for the population model and
then run parameterizing matrices.
}
\details{
This is an intermediate setup function to run the population, but some of its outputs are useful on their own, especially for eigen analyses. pop_model_setup() is run before pop_model_matrix_elements() and Projection_DD().
}
