% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationModel_Run.R
\name{PopulationModel_Run}
\alias{PopulationModel_Run}
\title{Population Model Run}
\usage{
PopulationModel_Run(
  dose = NA,
  sr_wb_dat = NA,
  life_cycle_params = NA,
  HUC_ID = NA,
  n_years = 100,
  MC_sims = 10,
  stressors = NA,
  output_type = "full"
)
}
\arguments{
\item{dose}{dataframe. Stressor magnitude dataset
imported from StressorMagnitudeWorkbook().}

\item{sr_wb_dat}{list object. Stressor response workbook
imported from StressorResponseWorkbook().}

\item{life_cycle_params}{dataframe. Life cycle parameters.}

\item{HUC_ID}{character. HUC_ID for the subwatershed unit.}

\item{n_years}{numeric. Number of years to run the population.}

\item{MC_sims}{numeric. set number of Monte
Carlo simulations for the Population Model.}

\item{stressors}{(optional) character vector of stressor
names to include in the Population Model. Leave the default
value as NA if you wish to include all stressors
applicable to the population model.}

\item{output_type}{(optional) character. Set to "full" for all data of "adults" for only adult data.}
}
\description{
Runs the population model
}
\details{
Runs the integrated population model within
the Joe Model stress-response framework. The population model is
run for a single subwatershed unit (HUC_ID). Inputs for the
PopulationModel_Run include the stressor-response and stressor magnitude
workbooks as well as the life cycle parameters table.
}
\examples{
\dontrun{
library(JoeModelCE)
}

}
