% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_CumulativeEffect.R,
%   R/Utility_PopulationModel.R
\name{proc}
\alias{proc}
\title{proc}
\usage{
proc(e, env = parent.frame())

proc(e, env = parent.frame())
}
\description{
proc

proc utility function
}
\keyword{internal}
