% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projection_DD.R
\name{Projection_DD}
\alias{Projection_DD}
\title{Project matrix with Density Dependence}
\usage{
Projection_DD(
  M.mx = NA,
  D.mx = NULL,
  H.mx = NULL,
  dat = NA,
  Nyears = 100,
  K = NA,
  p.cat = NA,
  CE_df = NULL
)
}
\arguments{
\item{M.mx}{A projection matrix expression}

\item{D.mx}{A matrix of density-dependence effect}

\item{H.mx}{A harm projection matrix}

\item{dat}{Life history data}

\item{Nyears}{Years to run simulation}

\item{K}{The population carrying Capacity}

\item{p.cat}{Probability of catastrophic event.}

\item{CE_df}{Cumulative effect dataframe. Data frame identifying cumulative effects stressors targets (system capacity or population parameter, or both, and target life stages.}
}
\value{
A list object with projected years, population size, lambda, fecundity, survival, catastrophic events.
}
\description{
Project the matrix model forward in time with density dependence.
}
\details{
Builds deterministic projection matrix using \code{popbio::stable.stage} with initial parameters based on arguments provided. Applies CE stressors to appropriate targets based on \code{CE_df}. All population modeling components are contained within this function. Users define a projection matrix, density-dependence matrix, harm projection matrix, life history parameters, years to simulate, carrying capacity, catastrophic event probabilities and a cumulative effects data frame (CE_df). When run this function will project the population forward in time. See the vignette tutorial Population Model Overview for details and instructions.
}
