% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StressorResponseWorkbook.R
\name{StressorResponseWorkbook}
\alias{StressorResponseWorkbook}
\title{Get Stressor Response Data from Excel Workbook}
\usage{
StressorResponseWorkbook(filename = NA)
}
\arguments{
\item{filename}{A string. Relative file name to the stressor response Excel workbook.}
}
\value{
A list object of the \code{main_sheet}, \code{stressor_names} and stressor response data frames (\code{sr_dat}).
}
\description{
Extract the stressor response data from each sheet in the Excel workbook.
}
\details{
For more information about data format requirements, see the Data Upload and Results Export tab of the associated Shiny App.

Uses \code{readxl::read_excel()} to load in workbook sheets:
\itemize{
\item Main sheet: meta data / cover sheet indicating:
\itemize{
\item stressor names
\item categories
\item interactions and linkages
\item scales
\item approximation function type
\item target life stages
\item contribution to survival or capacity parameters
\item stressor units
\item target model (CE or population models, or both)
}
\item all additional stressor-response sheets in workbook
\itemize{
\item context specific stressor-response curve data
\item sheet name must match stressor names in Main sheet, otherwise "Bad worksheet names" error message.
\item sheets must contain columns for \verb{<stressor_name>}, \verb{Mean System Capacity (\%)}, \code{SD}, \code{low.limit}, and \code{up.limit}
}
}
}
