% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemCapacity.R
\name{SystemCapacity}
\alias{SystemCapacity}
\title{System Capacity}
\usage{
SystemCapacity(
  f.dose.df,
  f.main.df,
  f.stressor.df,
  f.mean.resp.list,
  n.sims = 100
)
}
\arguments{
\item{f.dose.df}{Stressor magnitude data frame returned from \code{StressorMagnitudeWorkbook()} filtered for target stressor.}

\item{f.main.df}{\code{main_sheet} list object returned from \code{StressorResponseWorkbook()$main_sheet} filtered for target stressor.}

\item{f.stressor.df}{Dose response data frame returned from \code{StressorResponseWorkbook()$sr_dat} for target stressor.}

\item{f.mean.resp.list}{Response function list for target variable returned from \code{mean_Response()}.}

\item{n.sims}{Number of simulations to generate. Defaults t0 100.}
}
\description{
Calculates the system capacity for each stressor.
}
\details{
\code{SystemCapacity()} calculates the system capacity for each stressor. This function takes data frame arguments for stressor magnitudes and dose responses plus a list of stressor-response approximation functions (\code{f.} for local list or data frame). Note: Some stressors have multiple doses (Additive interaction).
}
